requirements.txt
# Map Categorizer Agent - Python 3.10+

# Google Maps
google-maps-services-python>=4.10.0

# HTTP & env
requests>=2.28.0
python-dotenv>=1.0.0

# Data
pandas>=2.0.0

# LLM (choose one or both)
openai>=1.0.0
anthropic>=0.18.0

# Optional: async / rate limiting
tenacity>=8.2.0
